libcamera: camera: Fix clearing of stream associations before `validate()`
A copy is made in the range-based for loop, and thus `setStream()`
operates on this copy, leading to the `stream_` member of the given
`StreamConfiguration` object in `*config` never being set to `nullptr`.

Fix that by taking a reference in the range-based for loop.

Also rename the variable from `it` since it is not an iterator.

Fixes: 4217c9f1aa863c ("libcamera: camera: Zero streams before validate()")
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>