libcamera: bitdepth: Add BitDepth implementation
Add a class template that simplifies bit depth conversion. This
functionality allows users to avoid having to perform inline bitshifting
to convert values of one bit depth to another.

For example, this makes it easier to input values from datasheets, where
a value may be expressed in a certain bit depth. It also removes the
potential for human errors when making these conversions manually, and
in a lot of cases makes bit depth conversions more readable.

Signed-off-by: Isaac Scott <isaac.scott@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>