pipeline: Add support for dumping capture script and metadata
Add support for dumping capture scripts and metadata using the Yaml
Emitter helpers.

The capture scripts can then be fed into the cam application and a
capture can thus be "replayed". Metadata can also be dumped.

Camera configuration is also dumped to the capture script. The cam
application currently does not support loading configuration from the
capture script, but support for that will be added in a subsequent
patch.

These can be enabled by a new environment variable.

Signed-off-by: Paul Elder <paul.elder@ideasonboard.com>
[rebased on the YamlEmitter classes]
Signed-off-by: Jacopo Mondi <jacopo.mondi@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>