libcamera: rpi: Add support for IMX258
Add support for the IMX258 camera. This patch adds a minimal cam_helper
implementation that allows the IMX258 sensor to be listed by cam, and
subsequently be used remotely in tools such as Camshark.

Based on the implementation of the IMX290's camera helper, adjusted to
use the equation listed in the IMX258 datasheet used to calculate
analogue gain.

Signed-off-by: Isaac Scott <isaac.scott@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>