android: Set StreamBuffer::srcBuffer in CameraDevice::processCaptureRequest
StreamBuffer::srcBuffer was set right before being processed, while it
was already determined when being constructed. This patch sets the value
in CameraDevice::processCaptureRequest. `requestedStreams` is refactored
from std::set to std::map to find the correct srcBuffer accordingly.

The change also helps cleaning up the post process pipeline's dependency
on `Camera3RequestDescriptor::pendingStreamsToProcess_`, which is going
to be cleaned up in the following patches.

Signed-off-by: Han-Lin Chen <hanlinchen@chromium.org>
Co-developed-by: Harvey Yang <chenghaoyang@chromium.org>
Signed-off-by: Harvey Yang <chenghaoyang@chromium.org>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>