libcamera: yaml_parser: Add `YamlObject::find(key)`
This function is meant to replace `YamlObject::contains()`
as it can be easily used for the same purpose, i.e. determining
if a certain key exists in a dictionary, but it has the advantage
that the result, if any, is immediately available for use
without having to do a second lookup.

Signed-off-by: Barnabás Pőcze <pobrn@protonmail.com>
Reviewed-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>