libcamera: converter_v4l2_m2m: Improve crop bounds support
When a converter (dw100 on imx8mp in this case) is used in a pipeline,
the ScalerCrop control get's created at createCamera() time. As this is
before configure(), no streams were configured on the converter and the
stream specific crop bounds are not known. Extend the converter class to
report the default crop bounds in case the provided stream is not found.

Signed-off-by: Stefan Klug <stefan.klug@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>