config: Introduce global runtime configuration
Currently, libcamera can be configured in runtime using several
environment variables.  With introducing more and more variables, this
mechanism reaches its limits.  It would be simpler and more flexible if
it was possible to configure libcamera in a single file.

For example, there have been a request for defining pipeline precedence
in runtime.  We want to compile in multiple pipelines, in order to have
them accessible within single packages in distributions.  And then being
able to select among the pipelines manually as needed based on the
particular hardware or operating system environment.  Having the
configuration file then allows easy switching between hardware, GPU or
CPU IPAs.  Another possible use case is tuning image output, especially
with software ISP, to user liking.  For example, some users may prefer
higher contrast without the need to use the corresponding knobs, if
present at all, in every application.  The configuration file can also
be used to enable or disable experimental features and avoid the need to
track local patches changing configuration options hard-wired in the
code when working on new features.

This patch introduces basic support for configuration files.  It reads,
stores and accesses the configuration.  The configuration is meant to
use as a hidden singleton with public access methods.  It is implemented
as GlobalConfiguration namespace exposing the public functions, with
private functions and the configuration data hidden in an anonymous
namespace.  This model suits well the purpose of having a hidden
singleton wrapped by public access functions.

libcamera configuration can be specified using a system-wide
configuration file or a user configuration file.  The user configuration
file takes precedence if present.  There is currently no way to merge
multiple configuration files, the one found is used as the only
configuration file.  If no configuration file is present, nothing
changes to the current libcamera behavior (except for some log
messages related to configuration file lookup).

The configuration file is a YAML file.  We already have a mechanism for
handling YAML configuration files in libcamera and the given
infrastructure can be reused for the purpose.  However, the
configuration type is abstracted to make contingent future change of the
underlying class easier while retaining (most of) the original API.

The configuration is versioned.  This has currently no particular
meaning but is likely to have its purpose in future, especially once
configuration validation is introduced.

The configuration YAML file looks as follows:

  ---
  version: 1
  configuration:
    WHATEVER CONFIGURATION NEEDED

There is no logging about reading the configuration file and contingent
errors.  This is on purpose because logging will query configuration,
which can lead to various problems when done during configuration
initialization.  Reporting the errors will be added later.

A complication arises from the fact that when the configuration is
loaded, logging may be called and logging will ask for the
configuration.  This is error-prone and may lead to subtle problems.
For this reason, the global configuration is instantiated to a pointer,
with an empty configuration initially.  The real configuration will be
created through initialize() method.  It will be clearer how it helps in
the followup patch introducing logging configuration.

Logging is also the most notable component from base that uses global
configuration.  In order to be able to do it, the global configuration
must be put to base.

This patch introduces just the basic idea.  Actually using the
configuration in the corresponding places (everything what is currently
configurable via environment variables should be configurable in the
file configuration) and other enhancements are implemented in the
followup patches.

Signed-off-by: Milan Zamazal <mzamazal@redhat.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>