ipa: rpi: Add a HW property to determine if the data buffer is strided
This property (cfeDataBufferStrided) indicates if the CSI-2 hardware
writes to the embedded/metadata buffer directly, or if it treats the
buffer like an image buffer and strides the metadata lines.

Unicam write this buffer strided, while the PiSP Frontend writes to it
directly. This information will be relevant to data parsers in the
helpers where the data is structured in lines.

Signed-off-by: Naushir Patuck <naush@raspberrypi.com>
Reviewed-by: David Plowman <david.plowman@raspberrypi.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>