treewide: Prefer `std::string_view` over `const std::string&` in parameters
A parameter of type `const std::string&` is not the right choice in
almost all cases. It forces the caller to construct an `std::string`
if they don't already have one, and accessing its data requires
an extra indirection. Furthermore, `std::string` is just one
instantiation of `std::basic_string<>`, using e.g. a different
allocator would immediately make it incompatible.

In contrast to that using an `std::string_view` is better if
NUL termination is not needed since no `std::string` construction
is required, and in many cases it can be passed directly in registers.

Signed-off-by: Barnabás Pőcze <pobrn@protonmail.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>