libcamera: base: log: Use std::string_view
Replace usage of const std::string references passed to class member
functions with std::string_view. This allows using static C string
literals in the callers without the overhead of constructing a
std::string instance.

As string views are not guaranteed to be zero-terminated, two usages
need to be modified:

- The call to syslog() is changed to use "%.*s" instead of "%s"

- The strtoul() call is replaced with std::from_chars()

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>