ipa: rkisp1: Initialize lsc tables on first frame
When sending the parameters for lens shading correction to the ISP on
frame x (in the test mostly frame 4) stalls of the ISP were observed.
From userland the symptom is that no buffers get returned anymore.
Kernel wise the ISP constantly produces interrupts with
RKISP1_CIF_ISP_V_START being set but no interrupts with
RKISP1_CIF_ISP_FRAME. It turned out that this behavior wasn't observed
when the lens shading correction parameters were written before
receiving the first frame. Ensure that behavior in the lsc algorithm by
unconditionally filling the params buffer on frame 0.

Unfortunately there were also cases were no stalls were observed even
though the lsc tables were not written on frame 0. So it is not clear
if this fix fixes the actual root cause.

Another reason for that change is that it is sensible to write the lsc
data on frame 0, to minimize the flicker in the image when the first
statistics come in (where they would have been written otherwise).

Signed-off-by: Stefan Klug <stefan.klug@ideasonboard.com>
Reviewed-by: Kieran Bingham <kieran.bingham@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>