libcamera: process: Pass stderr and reserve stdin and stdout fds
When a child process is started from Process::start(), the file
descriptors inherited from the parent process are closed, except
the ones explicitly listed in the fds[] argument.

One issue is that the file descriptors for stdin, stdout and stderr
being closed, the subsequent file descriptors created by the child
process will reuse the values 0, 1 and 2 that are now available.
Thus usage of printf(), assert() or alike may direct its output
to the new resource bound to one of these reused file descriptors.
The other issue is that the child process can no longer log on
the console because stderr has been closed.

To address the 2 issues, Process:start() is amended as below:
- Child process stdin and stdout are redirected to /dev/null so
that file descriptors 0 and 1 are not reused for any other
usage, that could be corrupted by the presence of printf(),
assert() or alike.
- Child process inherits from parent's stderr in order to share
the same logging descriptor.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>