libcamera: camera_sensor_properties: ov5675: Set correct delays
The OV5675 uses different delays for gain and exposure than are configured
in the default sensorDelays utilised by the CameraSensorLegacy.

Empirical testing using a Lenovo X13s shows that the exposure delay is
only a single frame, and the current setting of 2 frame delay produces
exceedingly frequent oscillations in the image exposure.

Update the OV5675 sensor delays table accordingly.

Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>