ipa: simple: Report the ColourGains in metadata
Provide the determined colour gains back into the metadata
to add to completed requests.

Metadata must be set before computing the new gain values in order to
report the metadata used to process the image rather than the metadata
determined from the image (and to be used for processing the next
image).

Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>