libipa: agc: Populate frameContext in queueRequest() in auto mode
In case of a param buffer underrun computeParams() is not called on a
frame. At the moment this causes a

"ERROR AgcMeanLuminance agc_mean_luminance.cpp:689 Effective exposure
value is 0. This is a bug in AGC and must be fixed for proper
operation."

log message, when processStats() is called on such a frame because
frameContext.gain is 0. Fix that by initializing all the values in
queueRequest().

Signed-off-by: Stefan Klug <stefan.klug@ideasonboard.com>
Reviewed-by: Paul Elder <paul.elder@ideasonboard.com>

---
Changes in v3:
- Manually rebased as agc was moved to libipa
- Changed commit message to better reflect the issue that this fixes

Changes in v2:
- Collected tag