apps: lc-compliance: Remove libevent dependency
libevent is only used as a way to notify the main thread from
the CameraManager thread when the capture should be terminated.

Not only is libevent not really necessary and instead can be
replaced with a simple combination of C++ STL parts, the current
usage is prone to race conditions.

Specifically, the camera's `queueRequest()` might complete the
request before the main thread could set the `loop_` member
variable and synchronize with the thread. This is a data race,
practically resulting in a nullptr or dangling pointer dereference.

This can easily be triggered by inserting a sufficiently large
timeout before `loop_ = new EventLoop;`:

  [46:43:40.529620351] [671833] DEBUG Request request.cpp:129 Request(4:C:0/1:0)
  ../src/apps/lc-compliance/helpers/capture.cpp:140:14: runtime error: member call on null pointer of type 'struct EventLoop'
    0 0x632c3e82f81a in CaptureBalanced::requestComplete(libcamera::Request*) ../src/apps/lc-compliance/helpers/capture.cpp:140

Signed-off-by: Barnabás Pőcze <pobrn@protonmail.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>