ipa: rkisp1: Add a lookahead of one frame when sending sensor controls
When the delayed controls instance of the pipeline is reset, it initializes
the values for frame 0 as being sent out to the sensor (which is
correct). The next sequence number that can be pushed to delayed
controls is therefore number 1. Ensure that the IPA never tries to
queue controls for frame 0.

Signed-off-by: Stefan Klug <stefan.klug@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>