libtuning: module: awb: Add bayes AWB support
To support the bayesian AWB algorithm in libtuning, the necessary data
needs to be collected and written to the tuning file.

Prior probabilities and AwbModes are defined by the user and therefore
added to the example config file.

Extend the output to also contain the necessary data for the bayesian
AWB algorithm.

Signed-off-by: Stefan Klug <stefan.klug@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>