apps: lc-compliance: Run request completion handler in "main" thread
Currently, `Capture::requestCompleted()` runs in the `CameraManager`'s
thread. This makes it a bit more complicated to use googletest and
report errors in those callbacks since lc-compliance sets up
googletest to throw exceptions on fatal errors / test skip, but
those exceptions are only caught on the "main" thread, the one
running the test suite.

To minimize the burden of dealing with synchronization in tests,
execute `Capture::requestCompleted()` in the event loop's thread
by utilizing `EventLoop::callLater()`.

Signed-off-by: Barnabás Pőcze <pobrn@protonmail.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>