libcamera: delayed_controls: Queue noop when needed, not before
A no-op is queued when DelayedControls runs out of controls at the end
of applyControls(). But these controls are only needed on the next call
to applyControls(), so there is still time for a proper push. To fix
that, move the no-op push to the beginning of applyContols().

Signed-off-by: Stefan Klug <stefan.klug@ideasonboard.com>

---

Changes in v3:
- Refrained dropping the writeCount_ member. It will be reused later

Changes in v2:
- Improved a log message
- Improved commit message
- dropped writeCount_ member