pipeline: rkisp1: Decouple image, stats and param buffers
The current code creates FrameInfo objects that tie together params,
stats and image buffers and expect that these always stay in these
groupings. However there are cases where these sequences get out of sync
(e.g. a scratch buffer is used for an image or a params buffer was too
late and therefore gets applied one frame later). As these situations
are timing related and cpu dependent it is impossible to guarantee the
initial grouping of buffers.

Split the buffers into separate queues for stats, images and params.
Resynchronize on image buffers as soon as they are dequeued from V4L2 as
these are the only buffers tied to the libcamera requests coming from
the user application.

Now handle the other buffer types according to their specific properties:

Stats buffers only need to be tracked after dequing and can be tied to
the corresponding request after the corresponding image buffer was
dequeued.

If params buffers get out of sync we can either inject the same set of
parameters twice or skip one set of params.

If image buffers get out of sync we need to update the expected sensor
sequence, so that the next request is assigned the correct sensor
sequence.

Signed-off-by: Stefan Klug <stefan.klug@ideasonboard.com>

---

Changes in v0.6:
- Fixed multiple assertions on start/stop and a few corner cases
- Fixed crash in dewarpRequestReady on stop

Changes in v0.5
- Fixed possible use-after-free in RequestInfo