ipa: libipa: Adjust for flicker in ExposureModeHelper
Update the ExposureModeHelper class to compensate for flickering
light sources in the ExposureModeHelper::splitExposure() function.
The adjustment simply caps exposure time at a multiple of the given
flicker period and compensates for any loss in the effective exposure
value by increasing analogue and then digital gain.

Initially in the one call-site for this function, a std::nullopt is
passed to make this a no-op.

Signed-off-by: Daniel Scally <dan.scally@ideasonboard.com>