ipa: libipa: Adjust for flicker in ExposureModeHelper
Update the ExposureModeHelper class to compensate for flickering
light sources in the ::splitExposure() function. The adjustment
simply caps exposure time at a multiple of the given flicker period
and compensates for any loss in the effective exposure value by
increasing analogue and then digital gain.

Initially in the one call-site for this function, a flicker period
of 0us is passed, making this a no-op.

Signed-off-by: Daniel Scally <dan.scally@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>