ipa: libipa: Add flicker controls to AgcMeanLuminance
Add controls for AeFlickerMode and AeFlickerPeriod to the
AgcMeanLuminance class. The controls passed to an algorithm's
queueRequest() are forwarded to AgcMeanLuminance through a new
function, and the values then passed to the ExposureModeHelper to be
used in calculating the new exposure time.

Take the opportunity to call the new parseControls() function in each
of the derived class' queueRequest() function.

Signed-off-by: Daniel Scally <dan.scally@ideasonboard.com>
Reviewed-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>